refactor: modularize foliage component
